requirements.txt
pycaret==3.0.1
pandas==1.5.3
flask==2.3.2
flask_httpauth==4.8.0
fastapi==0.100.0
uvicorn==0.23.1
pyyaml==6.0.1
requests==2.31.0
humanize==4.7.0
babel==2.12.1
sqlalchemy==1.4.49
pymysql==1.1.0
